+++ packages/core/forms/src/components/RedisConfigSelect.vue
@@ -121,8 +121,8 @@ const redisPartialFetcherKey: Ref<number, number> | undefined = inject(REDIS_PAR
 
 const endpoints = {
   konnect: {
-    getOne: '/v2/control-planes/{controlPlaneId}/core-entities/partials/{id}',
-    getAll: '/v2/control-planes/{controlPlaneId}/core-entities/partials',
+    getOne: '/v2/control-planes/{controlPlaneId}/core-entities/{workspace}/partials/{id}',
+    getAll: '/v2/control-planes/{controlPlaneId}/core-entities/{workspace}/partials',
   },
   kongManager: {
     getOne: '/{workspace}/partials/{id}',
@@ -213,11 +213,12 @@ const getOnePartialUrl = (partialId: string | number): string => {
 
   if (formConfig.app === 'konnect') {
     url = url.replace(/{controlPlaneId}/gi, formConfig?.controlPlaneId || '')
-  } else if (formConfig.app === 'kongManager') {
-    url = url.replace(/\/{workspace}/gi, formConfig?.workspace ? `/${formConfig.workspace}` : '')
   }
-  // Always replace the id when editing
-  url = url.replace(/{id}/gi, String(partialId))
+
+  url = url
+    .replace(/\/{workspace}/gi, formConfig?.workspace ? `/${formConfig.workspace}` : '')
+    .replace(/{id}/gi, String(partialId)) // Always replace the id when editing
+
   return url
 }