+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.3.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.3.27...@kong-ui-public/analytics-metric-provider@11.3.28) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.3.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.3.26...@kong-ui-public/analytics-metric-provider@11.3.27) (2026-05-10)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.3.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.3.25...@kong-ui-public/analytics-metric-provider@11.3.26) (2026-05-07)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider