+++ packages/entities/entities-consumers/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.4.3...@kong-ui-public/entities-consumers@4.4.4) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
+## [4.4.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.4.2...@kong-ui-public/entities-consumers@4.4.3) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
+## [4.4.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.4.1...@kong-ui-public/entities-consumers@4.4.2) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
+## [4.4.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.4.0...@kong-ui-public/entities-consumers@4.4.1) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumers
+
+
+
+
+
 # [4.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumers@4.3.32...@kong-ui-public/entities-consumers@4.4.0) (2026-05-08)