+++ packages/entities/entities-data-plane-nodes/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.36](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.6.35...@kong-ui-public/entities-data-plane-nodes@0.6.36) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes
+
+
+
+
+
+## [0.6.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.6.34...@kong-ui-public/entities-data-plane-nodes@0.6.35) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes
+
+
+
+
+
+## [0.6.34](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.6.33...@kong-ui-public/entities-data-plane-nodes@0.6.34) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes
+
+
+
+
+
+## [0.6.33](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.6.32...@kong-ui-public/entities-data-plane-nodes@0.6.33) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes
+
+
+
+
+
 ## [0.6.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-data-plane-nodes@0.6.31...@kong-ui-public/entities-data-plane-nodes@0.6.32) (2026-05-08)
 
 **Note:** Version bump only for package @kong-ui-public/entities-data-plane-nodes