+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.20.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.20.3...@kong-ui-public/entities-gateway-services@3.20.4) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.20.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.20.2...@kong-ui-public/entities-gateway-services@3.20.3) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.20.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.20.1...@kong-ui-public/entities-gateway-services@3.20.2) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.20.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.20.0...@kong-ui-public/entities-gateway-services@3.20.1) (2026-05-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
 # [3.20.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.19.17...@kong-ui-public/entities-gateway-services@3.20.0) (2026-05-08)