+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.46.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.46.0...@kong-ui-public/entities-shared@3.46.1) (2026-05-11)
+
+
+### Bug Fixes
+
+* **entities-shared:** z-index on deck command copy button ([#3242](https://github.com/Kong/public-ui-components/issues/3242)) ([8df058e](https://github.com/Kong/public-ui-components/commit/8df058e6c53b355583abf6226a80b0ab9d568b91))
+
+
+
+
+
+# [3.46.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.2...@kong-ui-public/entities-shared@3.46.0) (2026-05-11)
+
+
+### Features
+
+* **entities-shared:** add a secret advice to the decK customization editor ([#3240](https://github.com/Kong/public-ui-components/issues/3240)) ([ebfd53e](https://github.com/Kong/public-ui-components/commit/ebfd53e1ad7294890011ec54fc766fcdd025437e))
+
+
+
+
+
+## [3.45.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.1...@kong-ui-public/entities-shared@3.45.2) (2026-05-08)
+
+
+### Bug Fixes
+
+* **entities-consumer-groups:** query in exact match ([#3229](https://github.com/Kong/public-ui-components/issues/3229)) ([f9e5e43](https://github.com/Kong/public-ui-components/commit/f9e5e435adaad1af9c3fc91a737f36bece3e0496))
+
+
+
+
+
+## [3.45.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.0...@kong-ui-public/entities-shared@3.45.1) (2026-05-08)
+
+
+### Bug Fixes
+
+* **entities-shared:** use unredacted command for decK customization editor ([#3230](https://github.com/Kong/public-ui-components/issues/3230)) ([1f4ff15](https://github.com/Kong/public-ui-components/commit/1f4ff151adc7058690742ea76934a803d84a193b))
+
+
+
+
+
 # [3.45.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.44.4...@kong-ui-public/entities-shared@3.45.0) (2026-05-08)